requirements.txt
# requirements.txt - Python 3.13 compatible versions
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
google-generativeai==0.7.0
pydantic==2.9.0
pydantic-core==2.23.2
annotated-types==0.7.0
typing-extensions==4.15.0
starlette==0.38.6